+++ lib/reactotron-core-ui/src/contexts/CustomCommands/useCustomCommands.test.tsx
@@ -1,6 +1,6 @@
 /* eslint-disable react/display-name */
 import React from "react"
-import { act, renderHook } from "@testing-library/react-hooks"
+import { act, renderHook } from "@testing-library/react"
 
 import ReactotronContext from "../Reactotron"
 import { CommandType } from "reactotron-core-contract"